+++ app/smartbrain_3000/web/_app/immutable/chunks/Beb3yoZ1.js
@@ -1,0 +1,1 @@
+import"./CGwLJnGM.js";